+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.31
+
+* added `reset_and_start` to `measure`